requirements.txt
flask
requests
pandas
google-cloud-firestore
flask-login